Many changes here in implementing Redux. See below.
To start, had to run:
run npm install --save react-redux

And then run:
npm install --save redux

For rootReducer, I addd that to configureStore.js to keep things
easy.

In configureStore, the format of `import * as reducers from ./reducers`
did not work. So for now I simply import the testReducer alone. Will
have to fix and come up with way to combine the reducers effectively.

src/index.js now seems to be configured with the configureStore() function,
but to be tested.

BIG ERRORS here that took me a while to solve:
- From the redux instruction page, copying the language led to a "hidden"
character that could not be seen; only detected through arrow functions. Had
to go through each file where I copied and pasted and delete this hidden character.

- Also from the redux instruction page, the instructions (and code there)
for configStore.js say to call a function called composeWithDevtools
...BUT it is really composeWithDevTools.  Only noticed the error by chance
going through the actual docs.