IMPORTANT - On rendering a semantic UI React Grid...
...it seems that to really work, you cannot break up a single Grid
among multiple components.  I just can't get it to work. So each
compoenent that is part of a grid must be complete with <Grid> and
</Grid>.  This is turning out to be important because if I want to use
React Router and have it render multiple columns, each of which is
a separate component, I also need to wrap all of those "column" components
in a JSX tax.  This will not work:

<div>
<Grid.Row>
<Grid.Column>
Some text in Column 1
</Grid.Column>
</Grid.Row>

<Grid.Row>
<Grid.Column>
Some text in Column 2
</Grid.Column>
</Grid.Row>
</div>

If I do that, with the main <Grid> in some parent component, everything
gets screwed up.