IMPORTANT - Rework fetchBasicData in basicArtObjectActions to throw error...
...if there is no primaryimageurl.  Note how I made the catch almost like a
reducer -- it receives and object based on type.  So far, this was the only
way I could come up with/research to find a way to handle multiple errors
in this promise chain.  This has been tough to get to here.

I think this fixes the problem, though, with re-executing fetchBasicData.
To be tested further.