Rework currentArtObject reducer and helperActions#condenseRecord
Did this so that the reducer no long really defines the CAO state.
Now condenseRecord does. Also, deleted the case in the reducer dealing
with loading something and cleaned up some more of the keys of
COA to make it consistent throughout the app in preparation for connecting
this to DB.