Potentially resolve issue dealing with 422 error on POST to api server
Two things I had to do.  First was to add the header to the fetch request:

      headers: { "Content-Type": "application/json" },

Second was to add this to the api controller (art_objects_controller):

  skip_before_action :verify_authenticity_token

The key to solving this was to run the rails server on it's own and then
find the error in the output:

Started POST "/api/artobjects" for 127.0.0.1 at 2018-08-01 16:24:37 -0400
Processing by ArtObjectsController#create as */*
  Parameters: {"age"=>42, "art_object"=>{}}
Can't verify CSRF token authenticity.
Completed 422 Unprocessable Entity in 1ms (ActiveRecord: 0.0ms)

See here for big solution/discussion:

https://stackoverflow.com/questions/37269808/react-js-uncaught-in-promise-syntaxerror-unexpected-token-in-json-at-posit

https://stackoverflow.com/questions/35181340/rails-cant-verify-csrf-token-authenticity-when-making-a-post-request

ActionController::InvalidAuthenticityToken (ActionController::InvalidAuthenticityToken):