ON DEBUGGING and new class method self.find_and_update_lastViewed
For debugging, the problem with the model was I had to restart the server
completely.  Added new class method to find record and update last viewed,
because when I want to pull records from the last 30 days, I want that metric.
Maybe I could have used the updated_at field, but this seems less buggy.