IMP - Add postInitialObjectData to the actions that get artist, date and century art...
So now a user can click the buttons to find an artist, date, or culture and the
object is saved in the db and the id is added to the store/state!

This required fixing a problem with the new method in the ArtObject model,
find_and_update_lastViewed.  The key was returning nil if no record was found.