Abandoned attempt in last log. Too many issues.
So will leave repeated code in the components w/r/t
historyLinkClicked, removeFrom favorites, and addToFavorites.
The issue is the links wouldn't work unless i wrapped
the connect line in withRouter.  See here:

https://github.com/ReactTraining/react-router/issues/4105

Another problem though, is that these functions are a redirect
which requires access to this.props.history, and the parent Layout,
which renders the Routes, couldn't access the children's histories
to make the functions work.