Working on production env error - add home route and home action in...
...the art_object controller. Running the server in production mode
shows in the browser the same error that heroku is showing.  After some
research this seems to be an error (perhaps) related to not having
a "root" route in config/routes.  Running the server in production mode
now at least gets the rails default page that this is a rails app.

Command to run server in production mode:
rails server -e production