NOW WORKS ON HEROKU. The problem was...
...basically the same as here: https://github.com/AymaneZizi/Weatherpro/issues/1

The logs for browseum did not seem to be showing this however.  It only
kept popping up in the logs when I did heroku apps:create, heroku would
then create an app with a different name, and I would look at the build logs
for this seemingly new app.

I think also inserting the react-router stuff per the blog helped.

Now the problem is that the harvard api isn't resulting in anything.
I suspect this is an issue of react not having access to my key and me
needing to set my key in some kind of environment variable.